[data] [base-spells] Update spell data for dazzle and burn
Burn requires any moon or sun to be up to cast
Dazzle requires bright moon or sun to be cast

Burn requires any moon or sun to be up to cast
Dazzle requires bright moon or sun to be cast

This is a part 2 of 3 changes to enable combat-trainer to cast burn/dazzle when it can, and another spell when it can't. Needs to be 3 changes due to how dr-scripts updates work.

Part 1 - new common methods (this change)
Part 2 - new data fields in base-spells on the two spells
Part 3 - update to combat-trainer.