sort-scrolls - Added unleash option
Added option to specify a scroll to pull out, cast unleash on it, and then cast the spell if it was successful.

A waggle set for unleash is needed for this.  It needs to be set as follows, but you can edit how much mana you want to cast unleash at:

```
waggle_sets:
  unleash:
    Unleash:
      mana: 15
      cast: cast scroll
```
For this functionality I added 2 new yaml settings, `unleash_mana` and `unleash_harnesses`.  They are both integers and `unleash_mana` is the total mana you want to cast the unleased spell at and `unleash_harnesses` is how many times you want to harness to hold that much mana for the cast.

For example, if you set `unleash_mana` to 100 and then cast unleash on a scroll, it will perceive to see how much mana the spell is prepped at, take 100 - that, and then harness the rest by harnessing `unleash_harnesses` times.

I also added the option to specify the  spell by abbreviation so instead of this:
`;sort-scrolls find "whispers of the muse"'
You can now do this:
`;sort-scrolls find wotm`