[script][pick] Added defence against a timing issue where partial/additional traps are not recognised. (#7138)
Added waitrt? to a few locations to give time for additional traps to be
recognised by the code.

Should have no normal impact other than a fractional increase in time to
run the script. Hopefully the waitrt? will allow Flags['more-traps'] to
trip in time for the code to mark a partially trapped box as still
trapped (rather than letting it go through as untrapped, causing death
chaos and misery)

Note: this timing issue is sporadic and does not consistently present
itself.