[script] [stabbity] switch weapons if current is ineffective; avoid false-positives by not ignoring monsters (#7150)
### Background

While running the Arena on my thief in Duskruin this year, identified a
couple issues.

**Issue 1**
stabbity's yaml config lets you choose which weapon to use against
different creatures to handle when one might be resistant to slice or
impact, for example. However, this requires you to know the noun of the
creature before you encounter it. When progressing through the arena,
you might encounter new creatures that you haven't configured the
correct weapon for. To resolve, you have to update your yaml and rerun
the script.

**Issue 2**
stabbity identifies monsters in the room to attack based on NPCs and
removing any ignored NPCs (e.g. 'warrior' to ignore an Empath's Alfar
warrior). However, this can be a false-positive that ignores 'armored
necrotic warrior' leaving the script to do nothing.

### Changes

- Swap between preferred and alternate weapon if the current weapon is
not hitting effectively
- Don't ignore any NPCs so always try to watch and engage whoever is in
the arena