Added Learned Ranks tracking to exp-monitor.lic (#7146)
With the switchover to lich5, Wrayth/Stormfront users lost the function
to track EXP/RANKS gained over a session.

I have taken the portion of the script that handles tracking and
displaying the learned ranks in the field experience window.

Also, I have updated the output to Field Experience Window display to
not increase the number of characters, respacing the information to fit
in the original space.

I have included a screenshot of the before and after.


![expmoniter_learned_ranks](https://github.com/user-attachments/assets/419b92ef-38ed-47fa-900c-3b11eb9b6b4f)